src/main/java/br/com/damsete/domain/page/PageResponse.java
package br.com.damsete.domain.page;

import org.apache.commons.lang3.builder.ToStringBuilder;
import org.apache.commons.lang3.builder.ToStringStyle;

import java.util.List;

public class PageResponse<T> {

    private final List<T> content;
    private final int number;
    private final int size;
    private final int numberOfElements;
    private final int totalPages;
    private final long totalElements;
    private final boolean first;
    private final boolean last;

    public PageResponse(List<T> content, int number, int size, int numberOfElements, int totalPages,
                        long totalElements, boolean first, boolean last) {
        this.content = content;
        this.number = number;
        this.size = size;
        this.numberOfElements = numberOfElements;
        this.totalPages = totalPages;
        this.totalElements = totalElements;
        this.first = first;
        this.last = last;
    }

    public List<T> getContent() {
        return content;
    }

    public int getNumber() {
        return number;
    }

    public int getSize() {
        return size;
    }

    public int getNumberOfElements() {
        return numberOfElements;
    }

    public int getTotalPages() {
        return totalPages;
    }

    public long getTotalElements() {
        return totalElements;
    }

    public boolean isFirst() {
        return first;
    }

    public boolean isLast() {
        return last;
    }

    @Override
    public String toString() {
        return ToStringBuilder.reflectionToString(this, ToStringStyle.SHORT_PREFIX_STYLE);
    }
}
